Update patch set 1
Patch Set 1: Code-Review+2

Ravi, I am going to merge this commit to master, since it corrects at least some of the problems that are expected to exist with the configuration changes introduced with the rebasing into the policy/engine repo.

It is necessary though to tackle the comments from the code inspection with a new review.

Patch-set: 1
Label: Code-Review=+2