Update patch set 3
Patch Set 3: Code-Review+2

Thanks for the contribution Ruijing.   I gave it a test in mac os sierra.   I reduced number of CPUs  and memory resources to accommodate my system with less resources.   It installs OK using the latest version of vagrant 1.9.7.    I tried to give it a try too in an ubuntu LXC container with a brand new 16.04 image, and didn't quite work for me running the steps on the README.   I didn't spend time following up on the issues though, but related to virtualbox ..  I suppose the instructions may be a little more involved depending on the environment.

I suggest removing the instructions on the README.md and be more generic such having virtualbox and vagrant set up in there.    The commands and a discussion for the community where others can contribute could be done in a wiki page in onap, where other people can pitch in.

The Vagrantfile I see it more as a the template to use, but may need to be modified depending on particular system resources available.   Also note that the virtualbox configuration generated by vagrant may use all cpu resources or memory in the host system if not tweaked.   One idea, would be to chose those parameters, based on available system resources, may be from /proc subsystem.  Just a thought.   Other possibilities is just to to select which components to install programmatically, some may be interested in drools pdp but not in the UI ..

When login into the containers, I noticed that they are not properly set up, so although the installation and configuration completed ok, the containers where not configured completely.    Some error make sense as there is no DMaaP configured for example, others not sure ..

One question for you is what is the advantage to use vagrant vs for example doing what the setup_policy.sh is trying to do just on top of ubuntu and docker, removing the vagrant/virtualbox layer, just working directly with docker.

I assume this is work in progress, that you'll continue to enhance.

Patch-set: 3
Label: Code-Review=+2