Update patch set 1
Patch Set 1:

Ralph, there are 2 problems, the first one covered here, is that the featuresInstall and configure_component interfere at least reporting it messages with wrong output since both use a variable call name, so the configure_component.   Since I was fixing this, I just did it for the whole file, instead than for just one function.

The second one was related to the sed expression that expands variables fails in the generic configure_component function when there are special characters for example matching the delimiter, this will be hopefully corrected in my next commit shortly.

Patch-set: 1
Label: Code-Review=0