Update patch set 1
Patch Set 1:

I think this has to be merged under source control in order to work with "v1.1.1" images in lab environments.   The "v1.1.1" images uses "1.1.1" jar libraries.    This is part of the issue that we had in Marco's lab that he was using "v1.1.1" images.    There are no "1.1.2" in nexus/releases, so that was causing a problem.

Patch-set: 1
Reviewer: Gerrit User 477 <477@14b4e41f-c4e3-4fb9-9955-6a9b5656895a>
Label: Code-Review=0