Update patch set 1
Patch Set 1:

Moving forward, part of the issue is that there is no staging for amsterdam" source branches as in beijing "master" branch which it seems to use "staging" area.   The set up is different for amsterdam which complicates things a little bit, and the nature of this docker repo.

Probably root issue is to have to declare the version in this BRMS variable, if we get rid of that need, may be no issue to handle these situation.    May be something that we can give it shome thought and correct.

Patch-set: 1