workaround to insert groupentity tuple
This is a workaround for the issue described in POLICY-486.
The benefit of this workaround is that it does not require to
build new images, so it does not require a new load.

It takes advantage of the *tweaks.sh scripts, part of the "docker-compose"
project that is cloned for configuration previous to bring up the containers.

Change-Id: I91c7b2682bade0f3c244e37f960dd841c21a2e64
Issue-ID: POLICY-486
Signed-off-by: Jorge Hernandez <jh1730@att.com>