Update patch set 1
Patch Set 1:

> Alexis, Borislav, fyi, we'll need to update oom counterpart at some
 > point.   Borislav, you raised an issue in the past about
 > "pre-cooked data" in the image.  Hopefully this takes care of the
 > issue you experienced.

Thanks a lot! I raised a story https://jira.onap.org/browse/OOM-771 for this work.
This change will enable us to persist policy db.
As an effort to transition responsibility of maintaining deployment of components by teams themselves, would you like to do this change, raise a review and we will review and submit it?

Patch-set: 1