Update patch set 1
Patch Set 1:

Borislav, thanks. We will open a ticket to work with you folks for the updates, or assign that ticket OOM-771 to Jorge? It looks like OOM is currently still using Amsterdam v1.1.1 of Policy. Hopefully by M4 we will have Beijing release to begin testing with. We may need some guidance from you folks on this.

Patch-set: 1