Update patch set 1
Patch Set 1: Code-Review+1

It looks like we are changing nexus3 to align with dockerhub - no policy folder - good.  We will need to inform the community when this change is merged.  Also OOM will need the same change as heat in the values.yaml - I'll raise a bug on this work

Patch-set: 1
Label: Code-Review=+1