fix PE DB transaction lock timeouts.
I hope this fixes all DB related issues that
are being seen by integration teams.

The failOver jdbc parameters + perhaps some changes in underlying
code and persistence libraries may have been causing the
PAP to become useless.   A thread dump shows that a significant
number of threads were in timed-wait state, waiting on a lock
trying to acquire a connection.   In SB07, the system became
useless in terms wouldn't accept any transaction.

These changes to the PE provided jdbc parameters seems
to be behaving in the expected way, and the expectation is
that fixes the problems (behavior is good in own OOM lab).

It is thought that this may perhaps solve
POLICY-773, POLICY-775, POLICY-779, and perhaps POLICY-776.

Change-Id: Ib2319b68b5a503f61c2cb04998c549b7b97f9f31
Issue-ID: POLICY-777
Signed-off-by: Jorge Hernandez <jh1730@att.com>