No fetching applications load
drools-pdp in docker when it sees this configuration files,
it will try to execute them.   Since the alpine common
image has an immutable BUILD_VERSION that does not refer
to the actual drools build, it will not find it.

Two solutions: (a) update the scripts to look for the newly
added BUILD_VERSION_DROOLS to counter the presence of that
variable, or (b) remove them altogether since the control loop
application being the only one that is used will have the control
loop applications image cooked.

Change-Id: Ie8adb4b50565601b09e1100578458e3071ee6e54
Issue-ID: POLICY-1712
Signed-off-by: jhh <jorge.hernandez-herrero@att.com>