Update docker images for java 11
Added debian-slim images based off of jre-slim.
Removed alpine images from the build.

If it's determined that we need a jdk instead of a jre, then
"jre" can be changed to "jdk" in the policy-base Dockerfile.
As an alternative, a new policy-docker-jdk image could be created,
based off of "jdk" and all of the apt packages added to it.  In that
case, it would probably be appropriate to removed maven from the
packages that are loaded into the policy-common image.

Issue-ID: POLICY-1586
Signed-off-by: Jim Hahn <jrh3@att.com>
Change-Id: I43fd6a5c5ff9c4a0b537d3f297fdf2ad8248ba75