Update patch set 2
Patch Set 2:

(2 comments)

I tend to think that 2 root images are best.   One, lightweight for the use of api/pap/distribution/pdp-x as lighter weight components.   The openjdk jre over ubuntu buster slim is sufficient for it.

The other one, with the openjdk 11 + maven, heavier weight for the drools pdp, policy engine perhaps (as I think brmsgw needs it to deploy artifacts, sometimes maven in this image used to deal with maven deploy problems as well) , and perhaps apex.  I don't think there is a new for the policy-base one, I think it can easily be integrated into the other 2, it doesn't offer much benefit.

As I noted below, I would prefer to use the official maven image as the base for the second image, as it is official, and would be maintained by other folks, and we could just reuse.

Patch-set: 2
Label: Verified=0