Update patch set 2
Patch Set 2:

I think it is ok to keep snapshot version for now for the docker image (note that is the situation right now for drools-applications depending on the snapshot image from drools).   The concern is as we move along deriving from these base images on the policy components, we may want/need to do some tweaks in the base images.   I think it may save as releasing more than once if we wait until some basic tests are completed.

Patch-set: 2