Update patch set 2
Patch Set 2:

See commit text for background and intent.  I could propagate changes to policy-jre/debiam-slim to keep them consistent as not used in other images.  I I would like to do it in a different review.    Note that this is a prereq review for 3 others that tackle native policies, and broken CSITs affected by drools.   These are the dependent reviews that will be put out for inspection after this one in approved:

https://gerrit.onap.org/r/101802
https://gerrit.onap.org/r/101801
https://gerrit.onap.org/r/101803

Patch-set: 2