Update patch set 2
Patch Set 2:

Could we merge this one?   In next review in this module, I'll bump the versions, and do equivalent changes in the debian jre one.   Since the dependent images has not been modified to point to the released one, I think it is ok to proceed this way.   I would like to have the images merged to be able to work in the integration/csit repo changes, if you would agree.

Patch-set: 2
Reviewer: Gerrit User 477 <477@14b4e41f-c4e3-4fb9-9955-6a9b5656895a>
Label: Code-Review=0