Update patch set 2
Patch Set 2:

> Patch Set 2:
> 
> The FROM in pap,api, etc .. Dockerfiles will need to change to accommodate variable base image names, what's the plan for those components with regards to this change?

Hi,

We have a requirement to make proprietary builds of some selected ONAP components. These images are to be as close as possible to the opensource images but they are based on company specific base images, this is many to allow compliance with Company specific product rules. We have set up a build system that inserts the required Dockerfiles at the correct stage of the build so that we can use the proprietary base images. In order to stay as close as possible to the opensource we don't want to change the pom files in a proprietary way as that will put us on a branch that risks unneeded divergence. So this change allows us to build with our different images in a controlled way with minimal drift from the open source.

Let me know if you want more information.
/Andrew

Patch-set: 2
Reviewer: Gerrit User 249 <249@14b4e41f-c4e3-4fb9-9955-6a9b5656895a>