Merge csits into docker/csit
Decided to move CSIT test source code from the individual repos into
the docker repo.  This review modifies the scripts to begin the process,
and includes the CSIT code from xacml-pdp, as the first one to move.

Note: while the xacml-pdp files are new to the docker repo, they are
not actually new, thus there licenses were retained.

Issue-ID: POLICY-3299
Change-Id: I7cce548ceec38c9c1615af883f08eb118e84d132
Signed-off-by: Jim Hahn <jrh3@att.com>