Fix xacml-pdp csit test cases
In the recent runs, it was found that get statitcs might take few seconds longer after the actual deployment of policies. So, increasing the "max" wait time for the test cases to 1 min. And it won't slow down the tests as the check will be made every 15 sec to verify if the assertions are passing. The moment they pass, it will move on.

Issue-ID: POLICY-3257
Change-Id: Ifb00a7745b054fe490dd66cc680732893ff2b31a
Signed-off-by: Ram Krishna Verma <ram_krishna.verma@bell.ca>
(cherry picked from commit f1b994e85dd1075db23a4a23c7b2a3e88ac123c5)