Move remaining CSITs from to docker/csit
Moved the remaining CSITs from the individual policy repos to the
docker repo.  The files are unchanged, except for removing trailing
spaces and adding newlines to the end of a few of them.  The two
exceptions were the setup.sh scripts for drools-apps and distribution,
in which "/csit" is no longer being appended to the TESTPLANDIR, as
that variable now already includes "/csit".

Issue-ID: POLICY-3299
Change-Id: I4dd381ad7d916f334cfbcbfb541c78db97012dd2
Signed-off-by: Jim Hahn <jrh3@att.com>