Update APEX CSIT tests to cover multiple outputs
As part of this review, a new policy is executed in APEX-PDP that
interacts with AAI and CDS simulators. At each execution phase, along
with sending a request to AAI/CDS, a log event is also sent to DMaaP
notifying every action taken. This is done using the multiple output support.

Change-Id: If34df13823196f9ece9b5ce6448b6652507f0b63
Issue-ID: POLICY-3466
Signed-off-by: a.sreekumar <ajith.sreekumar@bell.ca>