Use jdk 11.0.11_9 in docker images
The onap java image only supports the java jre, but some of the
policy components need the java jdk.  Unfortunately, using alpine
"apk" to install openjdk11 does not result in the latest java jdk
version.

Copied the onap-jre Dockerfile contents and modified it to specify
the adopt jdk instead of the adopt jre.

Issue-ID: POLICY-3486
Change-Id: Icf833f1b1a068842d2141a84b5b3fe9aa1017b80
Signed-off-by: Jim Hahn <jrh3@att.com>