Base policy-jdk image on onap-python image
The onap-java image only installs the java jre, but policy needs the
jdk, thus we have been uninstalling the jre and then installing the jdk.
Unfortunately, this means that the policy-jdk image does not use the
latest ONAP-specified point release for either the java jdk OR python.

With this change, the policy-jdk image will start with the onap-python
image instead, thus python, at least will be up to date, even if the jdk
is not.

Note: tried using the latest adoptjdk image, but apparently there's
something different about that image that triggers a stack overflow in
apex-pdp.  Consequently, we'll stay with the openjdk installed via
alpine "apk".  This is still java 11, which satisfies ONAP requirements,
it just isn't the point release that is preferred.

Issue-ID: POLICY-3486
Change-Id: I0959c7048a38e30927b37b58deb887cc961a2150
Signed-off-by: Jim Hahn <jrh3@att.com>