Update patch set 1
Patch Set 1:

> Patch Set 1:
> 
> if an upgrade fails in the middle, how does this rollback start from the last successful script?  And how does it invoke the rollback/downgrade scripts in the reverse order?  (Similar question for downgrading.)

The upgrade function includes this block of code:
                if [ ${rc} -ne 0 ]; then
                    echo "${SCHEMA}: upgrade aborted at ${schemaVersion} by script ${sqlName}"
                    set_current_release "${schemaVersion}"
                    rollback "${PREVIOUS_RELEASE}" "${schemaVersion}" "${tag}"
                    return ${rc}
                fi
so if a script fails in the middle of an upgrade the rollback is triggered:
rollback "${PREVIOUS_RELEASE}" "${schemaVersion}" "${tag}"

this calls the rollback function which retrieves the list of scripts that were run successfully during the upgrade.
The history tables has a couple of new columns, id and tag.
id is a sequence number and tag groups upgrade/downgrade operations together.
tags are made up from date/time + version + single letter operation code (u/d/r)
The records (script names) are retrieved using the tag and are returned in reverse id order.
The names of the upgrade and rollback (downgrade scripts) are the same so it just runs the rollback versions. There is no rollback for downgrade at the moment.

Patch-set: 1