Update patch set 1
Patch Set 1:

> Patch Set 1:
> 
> the other review has been reverted and the corresponding distribution CSIT now passes, thus this review can be abandoned

Yes, I was just waiting on figuring out the APEX csit issue as well.
It seems the latest APEX-PDP image got built just an hour or so before the latest CDS upgrade changes. I did a remerge, and the APEX csits are passing too.
Abandoning this now.

Patch-set: 1