Update patch set 2
Patch Set 2:

This looks good Kevin, I think it is OK to leave the review open and wait for Jim to take a look.

Now that you proved that these changes can coexist with the jpa changes in api/pap as Ram indicated, I think we need to determine an strategy forward to do it.   A few things to consider:

- backup/restore strategy, should there be some form of it.

- moving forward, how to ensure that developers working in models don't forget to modify the sql tables as necessary.

- should clamp db tables be included ?

- plan for oom integration.

You can take a look at the job that does db configuration:

https://git.onap.org/oom/tree/kubernetes/policy/templates/job.yaml.

We would need db-migrator to run and integrate from this or a new job.

Other thoughts?

Patch-set: 2
Reviewer: Gerrit User 477 <477@14b4e41f-c4e3-4fb9-9955-6a9b5656895a>