Update patch set 1
Patch Set 1:

Renaming a bunch of files just to re-order (or insert/remove) items seems a bit tedious.  Maybe in Jakarta, we can re-think this strategy.  Perhaps we eliminate the prefixes and then add a single file, for each schema version, which lists the order of each script to be executed.  That same file would be used for both upgrades and downgrades.  I know we discussed this briefly during the design phase, but now the actual implementation is exposing the deficiencies in the current approach.

Patch-set: 1
Reviewer: Gerrit User 811 <811@14b4e41f-c4e3-4fb9-9955-6a9b5656895a>