Update patch set 1
Patch Set 1:

> Patch Set 1:
> 
> (1 comment)
> 
> I think it is best to keep the same

Hi Jorge. I am not sure of Liam's view on this as he is on leave today. There is an OOM review to change the https config - that will be cherry picked to Istanbul once approved. I thought it was his plan to raise a separate OOM review for the username but am not sure if he intended to cherry pick to Istanbul or not.

Patch-set: 1
Reviewer: Gerrit User 7603 <7603@14b4e41f-c4e3-4fb9-9955-6a9b5656895a>