Update patch set 1
Patch Set 1:

Thanks Ajith, I think you are the first one with this review doing actual upgrade/downgrade here.   I have a question about the actual data contents, let's say that you have the 2 new columns added in jakarta in jpatoscapolicy_target with data, and now for whatever reason you go to istanbul, the columns are dropped and the data goes away.    Should the data from those columns go somewhere else to maintain the integrity of the actual data for the istanbul layout.    May be the example is not that good but I think you get what I trying to say, we are modifying the layouts of the tables properly, but the question is if we are missing redistribution of the actual data contents (and not lose them)?    The question go in both directions upgrade and downgrade.

Patch-set: 1
Reviewer: Gerrit User 477 <477@14b4e41f-c4e3-4fb9-9955-6a9b5656895a>