Update patch set 1
Patch Set 1:

> Patch Set 1:
> 
> Thanks Ajith, I think you are the first one with this review doing actual upgrade/downgrade here.   I have a question about the actual data contents, let's say that you have the 2 new columns added in jakarta in jpatoscapolicy_target with data, and now for whatever reason you go to istanbul, the columns are dropped and the data goes away.    Should the data from those columns go somewhere else to maintain the integrity of the actual data for the istanbul layout.    May be the example is not that good but I think you get what I trying to say, we are modifying the layouts of the tables properly, but the question is if we are missing redistribution of the actual data contents (and not lose them)?    The question go in both directions upgrade and downgrade.

With the current implementation of the db-migrator, I think we don't give any consideration to the data in any of the tables. So, what happens today is,
say for e.g., there is 1 row in jpatoscapolicy_target "name1", "value1" under columns name, value.
When upgrading from I to J, two new columns are added: toscaPolicyName and toscaPolicyVersion. In the already existing rows, the 2 new columns will have empty values.
Similary on downgrading from J to I, any data  under columns toscaPolicyName and toscaPolicyVersion will be lost, and only the data under name and version will remain.

The tables affected in this review are not getting used really today, but I think your question is valid. If we want to save this data that is getting removed, then I think the question is how and where to save this data! I think "backup and restore" is the first thing we should have. Also, if we want to redistribute the data to a different table, maybe we could have scripts for that depending on how we want to do it.

Patch-set: 1