Update patch set 2
Patch Set 2: Code-Review-1

This is very nice although it introduces lots of duplications with respect to the csit directory.
Maintenance of the 2 directories of the duplicates is going to be difficult moving forward.
I suggest to extend the existing docker-compose-all.yml with the new components (we don't need to test them).
If that cannot be done for some reason, may be a docker-compose-prom.yml or something like that in the same directory with the csit and pointing to the same config files etc .. could be an alternative.

Patch-set: 2
Reviewer: Gerrit User 477 <477@14b4e41f-c4e3-4fb9-9955-6a9b5656895a>
Label: Code-Review=-1