Update patch set 2
Patch Set 2:

> Patch Set 2: Code-Review-1
> 
> This is very nice although it introduces lots of duplications with respect to the csit directory.
> Maintenance of the 2 directories of the duplicates is going to be difficult moving forward.
> I suggest to extend the existing docker-compose-all.yml with the new components (we don't need to test them).
> If that cannot be done for some reason, may be a docker-compose-prom.yml or something like that in the same directory with the csit and pointing to the same config files etc .. could be an alternative.

As I mentioned on the demo, the components need to start with the non https configuration, which is the default setting. So all the config files need to be set the http: false or useHttps: false, so not the same as csit. I can try to sort out some common folder, but that's the reason the the "duplicates", but they are not the same structure.

Patch-set: 2