Update patch set 7
Patch Set 7:

> Patch Set 7:
> 
> > Patch Set 7:
> > 
> > (2 comments)
> > 
> > thank you for consolidating it!
> 
> would be ok $VAR ?

My preference would be something that doesn't look like something else so it can be identified as a placeholder value.  I think the problem with $VAR is that it can be confused with a shell environment variable and provided that we are using bash scripts here, setting those will have no effect.   I used in the past {{X}} syntax but it can be anything that give it an unambiguous peculiar look.

Patch-set: 7
Reviewer: Gerrit User 477 <477@14b4e41f-c4e3-4fb9-9955-6a9b5656895a>