Fix issue with GeneratedValue in PfGeneratedIdKey
Until now, 'ID, name and version' was used as the primary key in
jpapolicyaudit and pdpstatistics tables. With this change, only ID will
be used as the PK.
The applications works with or without this change, as ID was the only
field which was always unique. "name" and "version" are mostly
duplicates (policyName in case of jpapolicyaudit & pdpName in case of
pdpstatistics table).
Pushing this change anyway to keep the tables consistent
with the entity classes.

Change-Id: I34a7582ed9c74fa0ead8fce2e86ac2ce1ecacf1e
Issue-ID: POLICY-3897
Signed-off-by: a.sreekumar <ajith.sreekumar@bell.ca>