Update patch set 2
Patch Set 2:

> Patch Set 1:
> 
> are we sure we want to change the database name, it breaks backwards compatibility from previous release?   This will also need corresponding oom changes.

I think it's OK for this one time in going from Istanbul to Jakarta for the TOSCA Control Loop/ACM features because we just have the PMSH use case as an installed base. We're handling that upgrade manually this time.

From Jakarta on, we will have to do this in the proper way.

Patch-set: 2
Reviewer: Gerrit User 2787 <2787@14b4e41f-c4e3-4fb9-9955-6a9b5656895a>