Update patch set 1
Patch Set 1:

When the currentInstanceCount is 0 and desiredInstanceCount is 1, in PAP CSIT, when consolidated healthcheck is done,the response says "healthy": false, as we dont bring up all the PDP instances in individual tests (e.g., in APEX csit, drools and xacml are not brought up).
As a result, https://github.com/onap/policy-docker/blob/master/csit/pap/tests/pap-test.robot#L39 this test fails.
3 options: 1)Either leave the currentInstanceCount as is, 2)update the consolidated healthcheck test in PAP and dont look for overall health 3) Bring up all pdps for all tests
I think either 1 or 2 options above are better, not 3. 
@Others: Please comment how to proceed.

Patch-set: 1