Update patch set 1
Patch Set 1:

> Patch Set 1:
> 
> When the currentInstanceCount is 0 and desiredInstanceCount is 1, in PAP CSIT, when consolidated healthcheck is done,the response says "healthy": false, as we dont bring up all the PDP instances in individual tests (e.g., in APEX csit, drools and xacml are not brought up).
> As a result, https://github.com/onap/policy-docker/blob/master/csit/pap/tests/pap-test.robot#L39 this test fails.
> 3 options: 1)Either leave the currentInstanceCount as is, 2)update the consolidated healthcheck test in PAP and dont look for overall health 3) Bring up all pdps for all tests
> I think either 1 or 2 options above are better, not 3. 
> @Others: Please comment how to proceed.

I think it would be good if we could have different groups.json for every csit test for each component that is tested and based only in the components that are brought up.    For example, if PAP tests does not bring up xacml and drools, those will be left out and healthcheck should pass.   In a similar manner if xacml tests doesn't need drools or apex, those will be out of it.   
I realized that this may be more work than what we want to do at this point but wanted to bounce it by you.   I would be OK too with #1 or #2 as you suggest if you decide those are best at this time.

Patch-set: 1
Reviewer: Gerrit User 477 <477@14b4e41f-c4e3-4fb9-9955-6a9b5656895a>