Update patch set 1
Patch Set 1:

> Patch Set 1:
> 
> I've been trying to debug pap csit test that is failing, but it runs just fine on local environmnent.

It seem that there is something that got broken in the CSITs lately.   If you look at the console output for 157 job:  https://jenkins.onap.org/view/policy/job/policy-pap-master-project-csit-verify-pap/157/console (last good one on June, 15th), policy-apex-pdp does not show up in docker-compose while it shows in this output (job 160).   It seems that this review from Wayne introduced the apex-pdp for the pap tests:  https://gerrit.onap.org/r/c/policy/docker/+/129623.   I suspect that there must be something in that review that breaks the CSITs.   Unfortunately, the verify and merge of that review, indicated that there was no problems, but all seems to point out that the issues you are hitting has to do with the changes introduced there, that touches the pap.

Patch-set: 1
Reviewer: Gerrit User 477 <477@14b4e41f-c4e3-4fb9-9955-6a9b5656895a>