Update patch set 1
Patch Set 1:

> Patch Set 1:
> 
> These scripts are failing because the CSITs are called in the verify job and they are failing due to the LF clearing down old snapshots.


I think we need to modify this part of script https://github.com/onap/policy-docker/blob/master/csit/get-versions.sh#L38 where we get the version of the docker image to pull from nexus. For example, for PAP istanbul version is 2.5.3-SNAPSHOT in pom.xml, this is processed as 2.5-SNAPSHOT-latest by the script. This should happen only for master branch, and for istanbul/jakarta etc, the version should be identified as the released version which is 2.5.2 which will be available in nexus.

Patch-set: 1
Reviewer: Gerrit User 4903 <4903@14b4e41f-c4e3-4fb9-9955-6a9b5656895a>