Update patch set 1
Patch Set 1:

> Patch Set 1:
> 
> > Patch Set 1:
> > 
> > These scripts are failing because the CSITs are called in the verify job and they are failing due to the LF clearing down old snapshots.
> 
> 
> I think we need to modify this part of script https://github.com/onap/policy-docker/blob/master/csit/get-versions.sh#L38 where we get the version of the docker image to pull from nexus. For example, for PAP istanbul version is 2.5.3-SNAPSHOT in pom.xml, this is processed as 2.5-SNAPSHOT-latest by the script. This should happen only for master branch, and for istanbul/jakarta etc, the version should be identified as the released version which is 2.5.2 which will be available in nexus.

Great minds think alike.......

I am working on a new version of that script that reads the release data in each branch and then checks in nexus to see which docker image is available. If the snapshot version is available it is used, otherwise the script steps back to the appropriate released version.

The input data for the scripts is those pf_release_data.csv files that we committed to the various branches 2 weeks ago.

Patch-set: 1
Reviewer: Gerrit User 2787 <2787@14b4e41f-c4e3-4fb9-9955-6a9b5656895a>