Update get version script for snapshots
The current script assumes that snapshot versions are always available.
However the Linux Foundation has now implemented a policy of deleting
snapshots on no-master branches that are older than a certain age. THis
breaks our CSITs.

THis updated script copes with this by checking if a snapshot is
avaialable and, if the snapshot is not available, the script falls back
to the released version.

This script will be needed on other branches besides master as well.

Issue-ID: POLICY-4233
Change-Id: I7a9fe7ffc3abfc8044a2e0439e3fcd47c086d5f8
Signed-off-by: liamfallon <liam.fallon@est.tech>