Update detmVers script for snapshots
The current script assumes that snapshot versions are always available.
However the Linux Foundation has now implemented a policy of deleting
snapshots on no-master branches that are older than a certain age. THis
breaks our CSITs.

THis updated script copes with this by checking if a snapshot is
avaialable and, if the snapshot is not available, the script falls back
to the released version.

This script will be needed on other branches besides master as well.

Issue-ID: POLICY-4233
Change-Id: I83f0a6816942d5bfaef7c22fc69625f40ae74d23
Signed-off-by: liamfallon <liam.fallon@est.tech>