Update patch set 13
Patch Set 13:

@liam - 
I think, I'll need you to regenarate the certificate stores that you produced.   I saw errors in previous runs where certificate issues when going to dmaap  (the simulator).   
I tried to grab a recent set of OOM stores to get new certs, but they are not doing the trick.   
Then I noticed that in the certificates that you added for the other components:  ks.jks and policy-trustore, they are different than the OOM ones, they have extra stuff and a different root ca.
One particular difference of the ks.jks is that the SAN in this keystore is missing an entry for
DNS Name: policy.api.simpledemo.onap.org
which is present in the original keystore and it is the hostname that docker-compose gives it to the simulator which does the dmaap that was giving trouble.   
If you could copy the stores, and place them into csit/config/drools-apps/custom/ as well, hopefully we could get a little further ?

Patch-set: 13
Reviewer: Gerrit User 477 <477@14b4e41f-c4e3-4fb9-9955-6a9b5656895a>