Ensure robot framework is installed for CSITs
The script to install the robot framework for CSITs was not called due
to an incorrect path in its invocation. This commit ensures it is
called.

This bug has been in the CSIT scripts for a numebr of releases. It only
seems to have become an issue in the last few hours.

Issue-ID: POLICY-4338
Change-Id: I0c2c578b784732a4f2beed514512f8e8c558f726
Signed-off-by: liamfallon <liam.fallon@est.tech>