Update patch set 6
Patch Set 6:

> Patch Set 6:
> 
> > Patch Set 6:
> > 
> > > Patch Set 6: Code-Review+1
> > > 
> > > I have concerns for maintenance, as the codebase is very similar to the official OOM one and will probably change in tandem.
> > 
> > Hi Jorge, Here we are not using any of the dependency charts like AAF, certInitializer, cert-wrapper, common etc and the policy charts can be deployed standalone in http. So the maintenance should be less complex comparing to OOM and also aids for testing policy components in kubernetes. The properties and image versions can be updated for every release as when required.
> 
> I might add that once the Kubernetes charts with automated CSIT work is finished and running reliably in Jenkins, we could remove the docker-compose related CSITs. This would mean that we do not need to maintain those anymore.
> 
> This would arguably mean less maintenance...

I think if the aim is to simply run the tests in a k8s environment instead of docker-compose I agree that once the job is completed, not many changes will be expected.   

If the aim is to try to replicate a somewhat and official OOM environment to give more assurance that things would work in an official installation, it would be more challenging maintenance I think as naturally OOM would drift apart from our charts as for examples they change in OOM for example with the shared mariadb, istio, strimzi. 

Another consideration is execution time to bring up the CSIT tests, wondering about the difference in time.

I think it may be be tough to beat the service that docker-compose CSITs have provided for tests and individual development over time but let's try it out and we'll see.

Patch-set: 6
Reviewer: Gerrit User 477 <477@14b4e41f-c4e3-4fb9-9955-6a9b5656895a>