Update patch set 1
Patch Set 1:

> Patch Set 1:
> 
> I just checked the console log for the apex-pdp tests, are there separate tests that run for Postgres? I couldn't see them in the log.

Hi Liam, I need to push an extra review to the ci-management repo to activate the apex-pdp CSITs that require postgres. The current apex tests are configured to use MariaDB. So, we need a separate job to set up the components that use postgres as their database to execute these tests. I have this review ready to push.

Patch-set: 1